+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+# Fixed
+
+- Typing on NumericStepper now works properly with `unitMultiplier` different than 1
+
 ## [9.135.3] - 2021-02-22
 
 ### Fixed